+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 ## Release Notes
 
+### TBD
+
+* Added `Path` object, `SwaggerApi#getPath(reqOrPath)` and `SwaggerApi#getPaths()`
+
 ### v0.1.0 (2015-08-12)
 
 * Initial release